requirements.txt
# Web Framework
fastapi==0.109.1
uvicorn==0.27.0
pydantic==2.5.2
python-multipart==0.0.7
python-dotenv==1.0.0
pydantic-settings

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL in production

# Twilio
twilio==8.10.0

# AI Models
openai==1.10.0
anthropic==0.8.0  # Optional alternative to OpenAI

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring and Observability
prometheus-client==0.17.1
sentry-sdk==1.35.0

# Utilities
tenacity==8.2.3  # For retries
python-json-logger==2.0.7
redis==5.0.1  # For caching (optional)

# Ngrok
pyngrok